+++ misc.cpp
@@ -24,6 +24,6 @@ void print_binary(uint32_t num){
 }
 
 //returns time in milliseconds
-double get_time(){
+uint64_t get_time(){
    return std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
 }
